API

API for GWinstekPSP. For full description of commands see users manual.

Index

GWinstekPSP.Status

GWinstekPSP.psp_decrementilimit

GWinstekPSP.psp_decrementplimit

GWinstekPSP.psp_decrementvlimit

GWinstekPSP.psp_decrementvoltage

GWinstekPSP.psp_incrementilimit

GWinstekPSP.psp_incrementplimit

GWinstekPSP.psp_incrementvlimit

GWinstekPSP.psp_incrementvoltage

GWinstekPSP.psp_readcurrent

GWinstekPSP.psp_readflags

GWinstekPSP.psp_readilimit

GWinstekPSP.psp_readplimit

GWinstekPSP.psp_readpower

GWinstekPSP.psp_readvlimit

GWinstekPSP.psp_readvoltage

GWinstekPSP.psp_savepreset

GWinstekPSP.psp_setilimit

GWinstekPSP.psp_setilimitmax

GWinstekPSP.psp_setknobfine

GWinstekPSP.psp_setknobnormal

GWinstekPSP.psp_setplimit

GWinstekPSP.psp_setplimitmax

GWinstekPSP.psp_setrelayoff

GWinstekPSP.psp_setrelayon

GWinstekPSP.psp_setvlimit

GWinstekPSP.psp_setvlimitmax

GWinstekPSP.psp_setvoltage

GWinstekPSP.psp_status

GWinstekPSP.psp_status!

GWinstekPSP.psp_togglerelay

Functions and Types

GWinstekPSP.Status — Type.

Power supply status

fields

voltage – Output voltage

current – Output current

power – Output power

vlimit – Voltage limit

ilimit – Current limit

plimit – Power limit

isrelayon – True if output relay is on

isoverheat – True if power supply temperature is too high

isknobfine – True if knob is in fine mode

isknobunlock– True if knob is unlocked

isremote – True if power supply is in remote control mode

islock – True if front panel controls are locked

source

GWinstekPSP.psp_status! — Function.

psp_status!(io, status)

Read status of the power supply and update status.

source

GWinstekPSP.psp_status — Function.

psp_status!(io)

Read status of the power supply and return status.

source

GWinstekPSP.psp_readvoltage — Function.

psp_readvoltage(io)

Read the output voltage.

source

GWinstekPSP.psp_readcurrent — Function.

psp_readcurrent(io)

Read the output current.

source

GWinstekPSP.psp_readpower — Function.

psp_readpower(io)

Read the output power.

source

GWinstekPSP.psp_readvlimit — Function.

psp_readvlimit(io)

Read the voltage limit.

source

GWinstekPSP.psp_readilimit — Function.

psp_readilimit(io)

Read the output current limit.

source

GWinstekPSP.psp_readplimit — Function.

psp_readplimit(io)

Read the power limit.

source

GWinstekPSP.psp_readflags — Function.

psp_readflags(io)

Read the flags.

Returns the tuple (isrelayon, isoverheat, isknobfine, isknobunlock, isremote, islock)

source

GWinstekPSP.psp_incrementvoltage — Function.

psp_incrementvoltage(io)

Add one unit to the voltage setting.

source

GWinstekPSP.psp_decrementvoltage — Function.

psp_decrementvoltage(io)

Subtract one unit to the voltage setting.

source

GWinstekPSP.psp_incrementvlimit — Function.

psp_incrementvlimit(io)

Add one unit to the voltage limit.

source

GWinstekPSP.psp_decrementvlimit — Function.

psp_decrementvlimit(io)

Add one unit to the voltage limit.

source

GWinstekPSP.psp_incrementilimit — Function.

psp_incrementilimit(io)

Add one unit to the current limit.

source

GWinstekPSP.psp_decrementilimit — Function.

psp_decrementilimit(io)

Subtract one unit to the current limit.

source

GWinstekPSP.psp_incrementplimit — Function.

psp_incrementplimit(io)

Add one unit to the power limit.

source

GWinstekPSP.psp_decrementplimit — Function.

psp_decrementplimit(io)

Subtract one unit to the power limit.

source

GWinstekPSP.psp_setvlimitmax — Function.

psp_setvlimitmax(io)

Set voltage limit to maximum value.

source

GWinstekPSP.psp_setilimitmax — Function.

psp_setilimitmax(io)

Set current limit to maximum value.

source

GWinstekPSP.psp_setplimitmax — Function.

psp_setplimitmax(io)

Set power limit to maximum value.

source

GWinstekPSP.psp_setknobfine — Function.

psp_setknobfine(io)

Set knob to fine.

source

GWinstekPSP.psp_setknobnormal — Function.

psp_setknobnormal(io)

Set knob to normal.

source

GWinstekPSP.psp_togglerelay — Function.

psp_togglerelay(io)

Toggle output relay.

source

GWinstekPSP.psp_setrelayon — Function.

psp_setrelayon(io)

Set output relay on.

source

GWinstekPSP.psp_setrelayoff — Function.

psp_setrelayoff(io)

Set output relay off.

source

GWinstekPSP.psp_savepreset — Function.

psp_savepreset(io)

Save current configuration to EEPROM.

source

GWinstekPSP.psp_setvoltage — Function.

psp_setvoltage(io, voltage)

Set output voltage.

source

GWinstekPSP.psp_setvlimit — Function.

psp_setvlimit(io, voltage_limit)

Set the voltage limit.

source

GWinstekPSP.psp_setilimit — Function.

psp_setilimit(io, current_limit)

Set the current limit.

source

GWinstekPSP.psp_setplimit — Function.

psp_setplimit(io, power_limit)

Set the power limit.

source